+++ src/main/java/io/orchestrate/client/ItemKind.java
@@ -1,0 +1,14 @@
+package io.orchestrate.client;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+
+public enum ItemKind {
+    ITEM,
+    EVENT,
+    RELATIONSHIP;
+
+    @JsonCreator
+    public static ItemKind fromJson(String name) {
+        return ItemKind.valueOf(name.toUpperCase());
+    }
+}